Mockup DOMManager
