+++ main.js
@@ -6,7 +6,9 @@ import './data_source/saturn';
 import './data_source/uranus';
 import './data_source/venus';
 
-import './getDOMObjects';
+import './DOMObjects';
+import './DOMUpdater';
+import './DOMManager';
 
 const planetsData ={
   earth: {
@@ -37,46 +39,4 @@ const planetsData ={
     title: `Venus`,
     data: {...planetVenus}
   },
-}
-
-
-window.addEventListener('hashchange', function() {
-	const newHash = new URL(document.URL).hash
-  const currentPlanet = newHash.split('#')[1]
-
-  updateDOMPlanetInfo(currentPlanet)
-});
-
-const updateDOMPlanetInfo = (planet) => {
-  const planetInfo = planetsData[planet]
-
-  const title = document.getElementById('planet-name')
-  const tag = document.getElementById('tag-earth')
-  const principalImg = document.getElementById('planet-principal-image')
-  const secondarImg = document.getElementById('planet-second-image')
-
-
-  title.innerText = planetInfo.title
-  principalImg.src = planetInfo.data.surfaceGeology.images[0]
-  secondarImg.src = planetInfo.data.surfaceGeology.images[1]
-  tag.addClass('active')
-}
-
-
-// Thia for onload at Body
-// overviewContent, internalStructure, surfaceGeology
-
-const checkPlanet = () => {
-  const curentUrl = new URL(document.URL)
-  const currentHash = curentUrl.hash
-  const currentPlanet = currentHash.split('#')[1]
-
-  if(!currentPlanet) {
-    const URLToGo = new URL(document.URL)
-
-    URLToGo.hash = '#earth'
-    return document.location.href = URLToGo
-  }
-
-  updateDOMPlanetInfo(currentPlanet)
 }
